chore: remove dead code and unused dependencies
Remove 31 unused crate-level dependencies and 12 unused workspace-level
dependencies across tapa-core and fpga-runtime workspaces. Remove dead
C++ code in tapacc (unused functions, dead #ifdef blocks, unused using
declarations) and tapa-lib (unused invoke_in_new_process and fork path).
Remove unused html-webpack-loader from tapa-visualizer.

+++ tapacc/rewriter/type.h
@@ -43,11 +43,11 @@ inline bool IsTapaType(const clang::ParmVarDecl* param,
   return IsTapaType(param->getType(), type_name);
 }
 
-inline std::string ArrayNameAt(const std::string& name, int idx) {
+inline std::string ArrayNameAt(const std::string& name, int64_t idx) {
   return name + "[" + std::to_string(idx) + "]";
 }
 
-inline std::string GetArrayElem(const std::string& name, int idx) {
+inline std::string GetArrayElem(const std::string& name, int64_t idx) {
   return name + "_" + std::to_string(idx) + "_offset";
 }
